+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed Quick Order Copy/Paste and Upload validation mis-evaluating item quantities when a More for Less promotion is active (B2BTEAM-3592)
+
 ## [3.16.8] - 2026-06-29
 
 ### Fixed